more funcs

+++ srcs/ft_strlcat.c
@@ -6,13 +6,15 @@
 /*   By: cefuente <cefuente@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/08 09:45:03 by cefuente          #+#    #+#             */
-/*   Updated: 2023/11/08 09:46:57 by cefuente         ###   ########.fr       */
+/*   Updated: 2023/11/08 10:47:05 by cefuente         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlen(char *str)
+#include "libft.h"
+
+size_t	ft_strlen(char *str)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
@@ -20,12 +22,12 @@ unsigned int	ft_strlen(char *str)
 	return (i);
 }
 
-unsigned int	ft_strlcat(char *dest, char *src, unsigned int size)
+size_t	ft_strlcat(char *dest, char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	j;
-	unsigned int	dest_length;
-	unsigned int	src_length;
+	size_t	i;
+	size_t	j;
+	size_t	dest_length;
+	size_t	src_length;
 
 	dest_length = ft_strlen(dest);
 	src_length = ft_strlen(src);
